# contracts/openapi/control-plane.v3.yaml
openapi: 3.1.0
info:
  title: NeKiro Control Plane API
  version: 3.0.0
  description: >-
    Gateway-owned Northbound API. Invocation v2 returns transient Agent output
    on the invocation request and exposes only metadata through Ledger reads.
servers:
  - url: https://api.nekiro.dev
    description: NeKiro Gateway destination
paths:
...
  /v3/workspaces/{workspaceId}/invocations:
    post:
      operationId: invokeAgent
      summary: Invoke an installed Agent and return its transient result
      description: >-
        This POST is the only Northbound result channel. The request stream
        field and Accept header MUST agree. stream=false requires
        application/json or a compatible wildcard and returns one Invocation
        Result. stream=true requires text/event-stream and returns ordered
        Invocation Result Stream Event data values. A mismatch returns 406
        NOT_ACCEPTABLE. Once SSE is committed, failed, canceled, and timed-out
        outcomes are in-band terminal events. EOF without a terminal event is
        interrupted delivery. A JSON result and every 502, 503, or 504 response
        MUST repeat the exact invocationId, rootTaskId, and traceId from the
        Invocation context created for this request. Results are not persisted
        or replayed.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ResultAccept"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvokeAgentRequest"
      responses:
        "200":
          description: Complete JSON result or committed SSE result stream selected by request mode
          content:
            application/json:
              schema:
                $ref: ../schemas/invocation-result.v1.schema.json
            text/event-stream:
              schema:
                type: string
              x-sse-data-schema:
                $ref: ../schemas/invocation-result-stream-event.v1.schema.json
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "406":
          $ref: "#/components/responses/NotAcceptable"
        "409":
          $ref: "#/components/responses/Conflict"
        "502":
          $ref: "#/components/responses/AgentFailure"
        "503":
          $ref: "#/components/responses/Unavailable"
        "504":
          $ref: "#/components/responses/Timeout"
  /v3/invocations/{invocationId}:
    get:
      operationId: getInvocation
      summary: Read an Invocation and metadata-only Ledger facts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/InvocationId"
      responses:
        "200":
          description: Invocation projection and append-only facts without Agent input or result content
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [invocation, events]
                properties:
                  invocation:
                    $ref: "#/components/schemas/InvocationRecord"
                  events:
                    type: array
                    items:
                      $ref: ../schemas/invocation-event.v0.2.schema.json
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/DependencyUnavailable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >-
        Gateway-verified bearer credential mapped to one trusted caller ID.
        Public caller-ID headers are not an authentication source.
...
  parameters:
...
    WorkspaceId:
      in: path
      name: workspaceId
      required: true
      schema:
        $ref: ../schemas/common.v1.schema.json#/$defs/workspaceId
...
    InvocationId:
      in: path
      name: invocationId
      required: true
      schema:
        $ref: ../schemas/common.v1.schema.json#/$defs/invocationId
    ResultAccept:
      in: header
      name: Accept
      required: true
      description: >-
        application/json or a compatible wildcard for stream=false;
        text/event-stream for stream=true.
      schema:
        type: string
        minLength: 1
  responses:
...
    ResourceNotFound:
      description: Requested resource was not found
      x-platform-error-codes: [NOT_FOUND]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v2.schema.json
...
    DependencyUnavailable:
      description: A required platform dependency failed
      x-platform-error-codes: [DEPENDENCY_ERROR]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v2.schema.json
    BadRequest:
      description: Invalid request
      x-platform-error-codes: [VALIDATION_ERROR]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v2.schema.json
    Unauthenticated:
      description: Authentication is required
      x-platform-error-codes: [UNAUTHENTICATED]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v2.schema.json
    Forbidden:
      description: Installation, Agent state, or capability is not authorized
      x-platform-error-codes: [FORBIDDEN, AGENT_DISABLED, CAPABILITY_NOT_ALLOWED]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v2.schema.json
    NotFound:
      description: Requested Agent or Installation was not found
      x-platform-error-codes: [NOT_FOUND, AGENT_NOT_INSTALLED]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v2.schema.json
    NotAcceptable:
      description: Request result mode and Accept header do not agree
      x-platform-error-codes: [NOT_ACCEPTABLE]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v2.schema.json
    Conflict:
      description: Invocation was canceled or conflicts with current state before response commitment
      x-platform-error-codes: [CONFLICT, CANCELED]
      content:
        application/json:
          schema:
            $ref: ../schemas/platform-error.v2.schema.json
    AgentFailure:
      description: Agent execution or A2A protocol failed before response commitment
      x-platform-error-codes: [AGENT_EXECUTION_FAILED, A2A_PROTOCOL_ERROR]
      x-platform-error-correlation:
        source: created-invocation-context
        exactFields: [invocationId, rootTaskId, traceId]
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CorrelatedPlatformError"
    Unavailable:
      description: Route, Agent, or required dependency is unavailable
      x-platform-error-codes: [ROUTE_NOT_FOUND, AGENT_UNAVAILABLE, DEPENDENCY_ERROR]
      x-platform-error-correlation:
        source: created-invocation-context
        exactFields: [invocationId, rootTaskId, traceId]
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CorrelatedPlatformError"
    Timeout:
      description: Invocation deadline expired before response commitment
      x-platform-error-codes: [TIMEOUT]
      x-platform-error-correlation:
        source: created-invocation-context
        exactFields: [invocationId, rootTaskId, traceId]
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CorrelatedPlatformError"
  schemas:
...
    CorrelatedPlatformError:
      allOf:
        - $ref: ../schemas/platform-error.v2.schema.json
        - type: object
          required: [invocationId, rootTaskId, traceId]
...
    InvokeAgentRequest:
      type: object
      additionalProperties: false
      required: [agentId, capability, input, stream]
      properties:
        agentId:
          $ref: ../schemas/common.v1.schema.json#/$defs/agentId
        capability:
          $ref: ../schemas/common.v1.schema.json#/$defs/capabilityId
        input:
          $ref: ../schemas/common.v1.schema.json#/$defs/jsonObject
        stream:
          type: boolean
    InvocationRecord:
      type: object
      additionalProperties: false
      required: [invocationId, rootTaskId, traceId, caller, workspaceId, targetAgentId, agentCardVersion, capability, status, createdAt, updatedAt]
      properties:
        invocationId:
          $ref: ../schemas/common.v1.schema.json#/$defs/invocationId
        rootTaskId:
          $ref: ../schemas/common.v1.schema.json#/$defs/taskId
        parentInvocationId:
          $ref: ../schemas/common.v1.schema.json#/$defs/invocationId
        traceId:
          $ref: ../schemas/common.v1.schema.json#/$defs/traceId
        caller:
          $ref: ../schemas/common.v1.schema.json#/$defs/caller
        workspaceId:
          $ref: ../schemas/common.v1.schema.json#/$defs/workspaceId
        targetAgentId:
          $ref: ../schemas/common.v1.schema.json#/$defs/agentId
        agentCardVersion:
          $ref: ../schemas/common.v1.schema.json#/$defs/semver
        capability:
          $ref: ../schemas/common.v1.schema.json#/$defs/capabilityId
        status:
          enum: [pending, routing, running, succeeded, failed, canceled, timed_out]
        latencyMs:
          type: integer
          minimum: 0
        errorCode:
          $ref: ../schemas/platform-error.v2.schema.json#/properties/code
        createdAt:
          $ref: ../schemas/common.v1.schema.json#/$defs/dateTime
        updatedAt:
          $ref: ../schemas/common.v1.schema.json#/$defs/dateTime